# backend-fltshr/src/main/java/com/zychp/backendfltshr/controllers/chore/ManagerQueueChoreController.java
package com.zychp.backendfltshr.controllers.chore;

import com.zychp.backendfltshr.model.chore.queuechores.QueueChoreCDTO;
import com.zychp.backendfltshr.model.chore.queuechores.QueueChoreDTO;
import com.zychp.backendfltshr.services.QueueChoreService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping(path = "manager/chores")
@RequiredArgsConstructor
public class ManagerQueueChoreController {
    private final QueueChoreService queueChoreService;

    @GetMapping("/queuechores")
    ResponseEntity getQueueChores() {
        return ResponseEntity.ok(queueChoreService.getQueueChores());
    }

    @PostMapping("/queuechore/user/{userId}")
    ResponseEntity<QueueChoreDTO> createQueueChore(@PathVariable Long userId,
                                                   @RequestBody QueueChoreCDTO queueChoreCDTO) {
        return ResponseEntity.accepted().body(queueChoreService.createQueueChore(userId, queueChoreCDTO));
    }

    @DeleteMapping("/queuechore/{queueChoreId}/archive")
    ResponseEntity archiveQueueChore(@PathVariable Long queueChoreId) {
        queueChoreService.archiveQueueChore(queueChoreId);
        return ResponseEntity.accepted().build();
    }
}