# wiki/swagger.yml
swagger: "2.0"
info:
  description: "Api documentation for FltShr"
  version: "1.0.0"
  title: "FltShr_API"
host: "localhost:81"
schemes:
  - "http"
paths:
...
  /manager/chores/queuechore/user/{userId}:
    post:
      tags:
        - Queue_Chores
      summary: Create new QueueChore and assign to first User
      parameters:
        - name: userId
          in: path
          description: user Id
          type: integer
          required: true
        - in: body
          name: ExpenseCDTO
          description: Expense Creation DTO
          schema:
            $ref: "#/definitions/QueueChoreCDTO"
      responses:
        201:
          description: Created
          schema:
            type: object
            $ref: "#/definitions/QueueChoreDTO"
  /manager/chores/queuechore/{queueChoreId}/archive:
    delete:
      tags:
        - Queue_Chores
      summary: Archive QueueChore
      parameters:
        - name: queueChoreId
          in: path
          description: QueueChore Id
          type: integer
          required: true
      responses:
        202:
          description: "Archived"
...
  /manager/chores/frequentchore/user/{userId}/date/{dateAssigned}:
    post:
      tags:
        - Frequent_Chores
      summary: Create new FrequentChores and assign to User
      parameters:
        - name: userId
          in: path
          description: user Id
          type: integer
          required: true
        - name: dateAssigned
          in: path
          description: Assign Date String "yyyy-MM-dd hh:mm:ss"
          type: integer
          required: true
        - in: body
          name: FrequentChoreCDTO
          description: FrequentChore creation DTO
          schema:
            $ref: "#/definitions/FrequentChoreCDTO"
      responses:
        201:
          description: Created
          schema:
            type: object
            $ref: "#/definitions/FrequentChoreDTO"
  /manager/chores/frequentchore/{frequentChoreId}/archive:
    delete:
      tags:
        - Frequent_Chores
      summary: Archive FrequentChore
      parameters:
        - name: frequentChoreId
          in: path
          description: FrequentChore Id
          type: integer
          required: true
      responses:
        202:
          description: "Archived"
          schema:
            type: object
            $ref: "#/definitions/FrequentChoreDTO"
...
definitions:
  QueueChoreDTO:
    type:  object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
  QueueChoreCDTO:
    type:  object
    properties:
      name:
        type: string
      description:
        type: string
      archived:
        type: boolean
...
  FrequentChoreDTO:
    type:  object
    properties:
      id:
        type: integer
      name:
        type: string
      description:
        type: string
      active:
        type: boolean
      durationDays:
        type: integer
      freqencyDays:
        type: integer
  FrequentChoreCDTO:
    type:  object
    properties:
      name:
        type: string
      description:
        type: string
      durationDays:
        type: integer
      freqencyDays:
        type: integer
...